Stabilise the merged C1 Ultra / SST / dynamic-linking tree
Catch-all commit for the in-progress merge work that had accumulated
on the c1-ultra-sst branch but never been recorded. Brings the source,
public headers, configs and unit tests into a state where the build
links, all 142 enabled unit tests pass, and the test programs run
under Docker.

Public-API headers (src/include/simeng/):
 - Move OS / kernel symbols into the OS namespace and the corresponding
   OS/ sub-headers (Process, SimOS, SyscallHandler), and route remaining
   callers through them.
 - Add compatibility shim headers at the legacy paths
   (BranchPredictor.hh, GenericPredictor.hh, MemoryInterface.hh,
   FixedLatencyMemoryInterface.hh, FlatMemoryInterface.hh) that
   forward to the new branchpredictors/ and memory/ locations, so
   existing includes keep compiling during the migration.
 - Refresh ReorderBuffer / DispatchIssueUnit / Architecture /
   Instruction / ExceptionHandler / MicroDecoder / RegisterFileSet
   declarations to match the implementation changes below.
 - Add aarch64/helpers/load.hh and store.hh placeholder headers so the
   refactored Instruction_execute.cc can include them uniformly with
   the other helpers.

Library (src/lib/):
 - AArch64: rework InstructionMetadata, Instruction, MicroDecoder and
   the per-stage Instruction_{address,decode,execute}.cc files to fit
   the new Capstone API (predicate operands via op.pred.reg, updated
   register enums) and the C1 Ultra pipeline.
 - RISC-V: parallel updates to Instruction / InstructionMetadata /
   Instruction_{address,decode,execute}.cc.
 - Pipeline: update ReorderBuffer, DispatchIssueUnit, ExecuteUnit,
   FetchUnit and the three core models (emulation, inorder, outoforder)
   for the merged interfaces.
 - Memory: align FixedLatencyMemoryInterface and FlatMemoryInterface
   with the post-merge MemoryInterface contract.
 - OS: refresh FileDesc, MemRegion, Process, SimOS and SyscallHandler
   for the OS-namespace migration; tweak SpecialFileDirGen and
   RegisterValue for the new helpers.
 - Misc: ModelConfig (Vector-Length defaults, new RS validation),
   GenericPredictor (TAGE plumbing), Elf reader, tools/simeng/main.cc.

Build glue:
 - Top-level CMakeLists.txt, src/lib/CMakeLists.txt, test/unit and
   test/integration / test/regression CMakeLists.txt: pick up the
   moved sources, drop the deleted ones, and wire in the SST build.

Configs:
 - configs/c1_ultra_geekbench.yaml: full set of pipeline, RS, and
   latency parameters for running Geekbench 6 on the C1 Ultra model.

Tests (test/unit/):
 - Refresh the AArch64 / RISC-V Architecture, Instruction, and
   ExceptionHandler unit tests for the new APIs; cut the bulk of the
   RISC-V ExceptionHandlerTest down to what still applies. Update
   MockCore, the memory-interface tests, OSTest, ProcessTest,
   PageFrameAllocatorTest, MemRegionTest, the pipeline tests
   (ExecuteUnit, RenameUnit, ReorderBuffer), and SimpleMem so they
   compile and run cleanly against the new headers.

+++ src/include/simeng/pipeline/DispatchIssueUnit.hh
@@ -11,6 +11,7 @@
 #include "simeng/Instruction.hh"
 #include "simeng/pipeline/PipelineBuffer.hh"
 #include "simeng/pipeline/PortAllocator.hh"
+#include "simeng/RegisterFileSet.hh"
 
 namespace simeng {
 namespace pipeline {
@@ -101,7 +102,7 @@ class DispatchIssueUnit {
   uint64_t getPortBusyStalls() const;
 
   /** Retrieve the current sizes and capacities of the reservation stations*/
-  void getRSSizes(std::vector<uint64_t>&) const;
+  void getRSSizes(std::vector<uint32_t>&) const;
 
  private:
   /** A buffer of instructions to dispatch and read operands for. */
