+++ src/lib/OS/FileDesc.cc
@@ -18,7 +18,7 @@ FileDescArray::FileDescArray() {
 }
 
 void FileDescArray::validateVfd(int vfd) const {
-  if (vfd < 0 || vfd > MAX_FD_NUM) {
+  if (vfd < 0 || static_cast<uint64_t>(vfd) > MAX_FD_NUM) {
     std::cerr << "[SimEng:FileDescArray] Invalid virtual file descriptor: "
               << vfd << std::endl;
     std::exit(1);
